pyproject.toml
[project]
name = "s3-fleet-management"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "datadog>=0.52.1",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "torch>=2.11.0",
]
